mysite/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-rest-framework==0.1.0
djangorestframework==3.16.1
pillow==12.0.0
psycopg==3.2.13
psycopg-binary==3.2.13
python-dotenv==1.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
